api/requirements.txt
flask
mysql-connector==2.1.6
